Upgrade Dashboard ResourcesPageFragment to 500 lines extracting all items.
